Revert "refactor(test): move autocapture test activities to autocapture package"
This reverts commit 6bcb552d73862a2aad87e8349a4097f85e0ee0d5.
