+++ analytics/src/androidTest/AndroidManifest.xml
@@ -5,10 +5,10 @@
             android:name="com.mixpanel.android.mpmetrics.TestActivity"
             android:exported="false" />
         <activity
-            android:name="com.mixpanel.android.autocapture.XmlAutocaptureTestActivity"
+            android:name="com.mixpanel.android.mpmetrics.XmlAutocaptureTestActivity"
             android:exported="false" />
         <activity
-            android:name="com.mixpanel.android.autocapture.ComposeAutocaptureTestActivity"
+            android:name="com.mixpanel.android.mpmetrics.ComposeAutocaptureTestActivity"
             android:exported="false" />
     </application>
 </manifest>
