+++ analytics/src/androidTest/java/com/mixpanel/android/mpmetrics/ComposeAutocaptureTestActivity.kt
@@ -1,4 +1,4 @@
-package com.mixpanel.android.autocapture
+package com.mixpanel.android.mpmetrics
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
