+++ analytics/src/androidTest/java/com/mixpanel/android/mpmetrics/XmlAutocaptureInstrumentedTest.java
@@ -4,8 +4,6 @@ import android.content.Context;
 import android.content.SharedPreferences;
 import android.view.View;
 
-import com.mixpanel.android.autocapture.XmlAutocaptureTestActivity;
-
 import androidx.test.core.app.ActivityScenario;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.filters.LargeTest;
